+++ src/configSearch/criteria.js
@@ -1,278 +1,161 @@
+const Operator = {
+  "name": "Condition",
+  "key": "opCondition",
+  "items": [
+    {
+      "name": "equal"
+    },
+    {
+      "name": "not equal",
+    }
+  ]
+};
+
 export const CRITERIA = {
-    "categories" : {
-      "name": "Category",
-      "input": [
-        {
-          "name": "Condition",
-          "key": "opCondition",
-          "items": [
-            {
-              "name": "equal"
-            },
-            {
-              "name": "not equal",
-            }
-          ]
-        },
-        {
+  Category : [
+      Operator,
+      {
           "name": "List Category",
           "key": "list",
           "items": []
-        },
-        {
+      },
+      {
           "name": "Experince",
           "key": "experience",
           "items": [
-            {
+          {
               "name": "1"
-            },
-            {
+          },
+          {
               "name": "2",
-            },
-            {
+          },
+          {
               "name": "3"
-            },
-            {
+          },
+          {
               "name": "4",
-            },
-            {
+          },
+          {
               "name": "5"
-            },
-            {
+          },
+          {
               "name": "6"
-            },
-            {
+          },
+          {
               "name": "7",
-            },
-            {
+          },
+          {
               "name": "8"
-            },
-            {
+          },
+          {
               "name": "9",
-            },
-            {
+          },
+          {
               "name": "10"
-            }
+          }
           ]
-        },
-        {
+      },
+      {
           "name": "Proficiency",
           "key": "proficiency",
           "items": [
-            {
+          {
               "name": "1"
-            },
-            {
+          },
+          {
               "name": "2",
-            },
-            {
+          },
+          {
               "name": "3"
-            },
-            {
+          },
+          {
               "name": "4",
-            },
-            {
+          },
+          {
               "name": "5"
-            }
+          }
           ]
-        },
-
-      ]
-    },
-    "skills" : {
-      "name": "Skill",
-      "input": [
-        {
-          "name": "Condition",
-          "key": "opCondition",
-          "items": [
-            {
-              "name": "equal"
-            },
-            {
-              "name": "not equal",
-            }
-          ]
-        },
-        {
-          "name": "List Skills",
-          "key": "list",
-          "items": []
-        },
-        {
-          "name": "Experince",
-          "key": "experience",
-          "items": [
-            {
-              "name": "1"
-            },
-            {
-              "name": "2",
-            },
-            {
-              "name": "3"
-            },
-            {
-              "name": "4",
-            },
-            {
-              "name": "5"
-            },
-            {
-              "name": "6"
-            },
-            {
-              "name": "7",
-            },
-            {
-              "name": "8"
-            },
-            {
-              "name": "9",
-            },
-            {
-              "name": "10"
-            }
-          ]
-        },
-        {
-          "name": "Proficiency",
-          "key": "proficiency",
-          "items": [
-            {
-              "name": "1"
-            },
-            {
-              "name": "2",
-            },
-            {
-              "name": "3"
-            },
-            {
-              "name": "4",
-            },
-            {
-              "name": "5"
-            }
-          ]
-        },
-      ]
-  
-    },
-    "branch": {
-      "name": "Branch",
-      "input": [
-        {
-          "name": "Condition",
-          "key": "opCondition",
-          "items": [
-            {
-              "name": "equal"
-            },
-            {
-              "name": "not equal",
-            }
-          ]
-        },
-        {
-          "name": "Branch",
-          "key": "branch",
-          "items": [
-            {
-              "name": "Goris"
-            },
-            {
-              "name": "Gyumri"
-            },
-            {
-              "name": "Erevan"
-            },
-            {
-              "name": "Kapan"
-            },
-            {
-              "name": "Stepanakert"
-            },
-            {
-              "name": "Vanadzor"
-            }
-          ]
-        }
-      ]
-  
-    },
-    "position": {
-      "name": "Position",
-      "input": [
-        {
-          "name": "Condition",
-          "key": "opCondition",
-          "items": [
-            {
-              "name": "equal"
-            },
-            {
-              "name": "not equal",
-            }
-          ]
-        },
-        {
-          "name": "Position",
-          "key": "position",
-          "items": [
-            {
-              "name": "Beginner SW Engineer"
-            },
-            {
-              "name": "SW Engineer"
-            },
-            {
-              "name": "Senior SW Engineer"
-            },
-            {
-              "name": "Beginner QA Tester"
-            },
-            {
-              "name": "QA Tester"
-            },
-            {
-              "name": "SQE Analyst"
-            },
-            {
-              "name": "Sr. Software Quality Engineer"
-            },
-            {
-              "name": "QA Analyst"
-            },
-            {
-              "name": "QA lead"
-            },
-            {
-              "name": "Team lead"
-            },
-            {
-              "name": "Graphic designer"
-            },
-            {
-              "name": "Technical manager"
-            },
-            {
-              "name": "Senior Team lead"
-            },
-            {
-              "name": "Project Manager"
-            },
-            {
-              "name": "3D modeler"
-            },
-            {
-              "name": "UIUX designer"
-            },
-            {
-              "name": "SW Architect"
-            },
-          ]
-        },
-      ]
-    }
-  };
+      }
+  ],
+  Skill : [
+      Operator,
+      {
+        "name": "List Skills",
+        "key": "list",
+        "items": []
+      },
+      {
+        "name": "Experince",
+        "key": "experience",
+        "items": [
+          {
+            "name": "1"
+          },
+          {
+            "name": "2",
+          },
+          {
+            "name": "3"
+          },
+          {
+            "name": "4",
+          },
+          {
+            "name": "5"
+          },
+          {
+            "name": "6"
+          },
+          {
+            "name": "7",
+          },
+          {
+            "name": "8"
+          },
+          {
+            "name": "9",
+          },
+          {
+            "name": "10"
+          }
+        ]
+      },
+      {
+        "name": "Proficiency",
+        "key": "proficiency",
+        "items": [
+          {
+            "name": "1"
+          },
+          {
+            "name": "2",
+          },
+          {
+            "name": "3"
+          },
+          {
+            "name": "4",
+          },
+          {
+            "name": "5"
+          }
+        ]
+      },
+    ],
+  Branch:  [
+      Operator,
+      {
+        "name": "Branch",
+        "key": "branch",
+        "items": []
+      }
+    ],
+  Position: [
+      Operator,
+      {
+        "name": "Position",
+        "key": "position",
+        "items": []
+      },
+    ]
+};
