+++ src/store/actions/search.js
@@ -1,6 +1,7 @@
 import { get, post } from 'client/lib/axiosWrapper';
 import {searchUsersBegin, searchUsersSuccess, searchUsersFailure, criteriaRow} from './searchAction';
 import { CRITERIA } from '../../configSearch/criteria';
+import base64 from 'base-64';
 
 const id = (+ new Date() + Math.floor(Math.random() * 999999)).toString(36);
 
@@ -18,30 +19,28 @@ export function getSearchParams(params){
 
 const getIdValues = (type, item) => {
   let id;
-  Object.values(CRITERIA).map((criteria) => {
-      if(criteria.name !== type) {
+  Object.keys(CRITERIA).map((criteria) => {
+      if(criteria !== type) {
           return null;
       }
 
-      criteria.input.map((data, index) => {
-          if(data.key === "list"){
+      Object.values(CRITERIA[criteria]).map((data, index) => {
               data.items.map(e => {
                   if(e.name === item){
                       id = e.id;
                   }
               })
-          }
       });
   });
   return id;
   }
 
 function getCriteria(){
-  Object.values(CRITERIA).map(field => {
-    if(field.name === "Category"){
+  Object.keys(CRITERIA).map(field => {
+    if(field === "Category"){
       get({url: "categories/all"}).then(result => {
-        result && result.data.map(item => {
-             Object.values(field.input).map(cat => {
+        result.data.map(item => {
+             Object.values(CRITERIA[field]).map(cat => {
                  if(cat.key === "list"){
                      cat.items.push({name: item.name, id: item.id});
                  }
@@ -52,11 +51,39 @@ function getCriteria(){
           //TODO handle error
           console.error("Error get categories: ", error);
       });
-    } else {
+    } else if(field === "Skills") {
       get({url: "skills"}).then(result => {
-        result && result.data.map(item => {
-             Object.values(field.input).map(cat => {
-                 if(cat.key === "list"){
+        result.data.map(item => {
+             Object.values(CRITERIA[field]).map(cat => {
+                 if(cat.key === "skills"){
+                     cat.items.push({name: item.name,  id: item.id});
+                 }
+             })
+         });
+      })
+      .catch(error => {
+          //TODO handle error
+          console.error("Error get categories: ", error);
+      });
+    } else if(field === "Branch") {
+      get({url: "branch"}).then(result => {
+        result.data.map(item => {
+             Object.values(CRITERIA[field]).map(cat => {
+                 if(cat.key === "branch"){
+                     cat.items.push({name: item.name,  id: item.id});
+                 }
+             })
+         });
+      })
+      .catch(error => {
+          //TODO handle error
+          console.error("Error get categories: ", error);
+      });
+    } else {
+      get({url: "position"}).then(result => {
+        result.data.map(item => {
+             Object.values(CRITERIA[field]).map(cat => {
+                 if(cat.key === "position"){
                      cat.items.push({name: item.name,  id: item.id});
                  }
              })
@@ -72,34 +99,14 @@ function getCriteria(){
 
 }
 
-//TODO: Move to useEffect
-// getCriteria();
+getCriteria();
 
 export function getUsers(data){
     return (dispatch) => {
         dispatch(searchUsersBegin());
-        const bodyObject = data && Object.values(data).map(item => {
-          const itemObject  = (
-              (item.type === "Branch") ?
-                  { name: item.branch } :
-              (item.type === "Position") ?
-                  { name: item.position } :
-              {
-                  id: getIdValues(item.type, item.list),
-                  // name: item.list,
-                  experience: item.experience,
-                  profficience: item.proficiency
-              }
-          );
 
-              return {
-                  type: item.type && item.type.toLowerCase(),
-                  opCondition: item.opCondition,
-                  items: itemObject,
-                  relCondition: item.relCondition ? item.relCondition.toLowerCase() : "and"
-              }
-          })
-        post({url: "search/", data: bodyObject})
+        const encData = base64.encode(JSON.stringify(data));
+        post({url: `search?=${encData}`, data: encData})
         .then(result => {
           dispatch(searchUsersSuccess({data: result.data.users, values: data, rows: searchParams}));
         }).catch(error => {
